file_server/requirements.txt
flask_cors
# Flask-SSLify
flask-sslify==1.0.0